test(di): isolate disposal ownership regressions
